+++ sentry-android-core/src/main/java/io/sentry/android/core/InternalSentrySdk.java
@@ -103,9 +103,7 @@ public final class InternalSentrySdk {
       app.setAppName(ContextUtils.getApplicationName(context, options.getLogger()));
 
       final @NotNull TimeSpan appStartTimeSpan =
-          options.isEnableStarfish()
-              ? AppStartMetrics.getInstance().getAppStartTimeSpan()
-              : AppStartMetrics.getInstance().getSdkAppStartTimeSpan();
+          AppStartMetrics.getInstance().getAppStartTimeSpanWithFallback();
       if (appStartTimeSpan.hasStarted()) {
         app.setAppStartTime(DateUtils.toUtilDate(appStartTimeSpan.getStartTimestamp()));
       }