+++ sentry-android-core/src/main/java/io/sentry/android/core/SentryAndroidOptions.java
@@ -212,7 +212,7 @@ public final class SentryAndroidOptions extends SentryOptions {
    */
   private boolean attachAnrThreadDump = false;
 
-  private boolean enableStarfish;
+  private boolean enablePerformanceV2 = false;
 
   public SentryAndroidOptions() {
     setSentryClientName(BuildConfig.SENTRY_ANDROID_SDK_NAME + "/" + BuildConfig.VERSION_NAME);
@@ -593,12 +593,12 @@ public final class SentryAndroidOptions extends SentryOptions {
   }
 
   @ApiStatus.Experimental
-  public boolean isEnableStarfish() {
-    return enableStarfish;
+  public boolean isEnablePerformanceV2() {
+    return enablePerformanceV2;
   }
 
   @ApiStatus.Experimental
-  public void setEnableStarfish(final boolean enableStarfish) {
-    this.enableStarfish = enableStarfish;
+  public void setEnablePerformanceV2(final boolean enablePerformanceV2) {
+    this.enablePerformanceV2 = enablePerformanceV2;
   }
 }