+++ sentry-android-core/src/test/java/io/sentry/android/core/SentryAndroidOptionsTest.kt
@@ -142,16 +142,16 @@ class SentryAndroidOptionsTest {
     }
 
     @Test
-    fun `starfish is disabled by default`() {
+    fun `performance v2 is disabled by default`() {
         val sentryOptions = SentryAndroidOptions()
-        assertFalse(sentryOptions.isEnableStarfish)
+        assertFalse(sentryOptions.isEnablePerformanceV2)
     }
 
     @Test
-    fun `starfish can be enabled`() {
+    fun `performance v2 can be enabled`() {
         val sentryOptions = SentryAndroidOptions()
-        sentryOptions.isEnableStarfish = true
-        assertTrue(sentryOptions.isEnableStarfish)
+        sentryOptions.isEnablePerformanceV2 = true
+        assertTrue(sentryOptions.isEnablePerformanceV2)
     }
 
     fun `when options is initialized, enableScopeSync is enabled by default`() {